Extend inference_pipe to support video inference & Update the build system (#8)
* Added predict_video function that takes in video and returns prediction array for each frame. Added a unit test for function.

* Migrates to setuptools build system

Switches from a requirements.txt based dependency management to using setuptools with pyproject.toml and setup.cfg.
This allows for better package management and supports editable installs with development dependencies.
Removes the model files from the repository and adds them to the .gitignore.
Updates the README with new installation instructions.

---------

Co-authored-by: Stealth <47306082+StealthHydra179@users.noreply.github.com>